Build types index on docker build
