+++ .dockerignore
@@ -1,0 +1,5 @@
+Dockerfile
+.git
+.dockerfile
+types.idx
+types.json
